backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
google-cloud-aiplatform==1.38.0
google-generativeai==0.3.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-maps-services==4.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
requests==2.31.0
geopy==2.4.1
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
gunicorn==21.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0